Add PackageManagerClient API to intiate download and install of a package
