+++ src/DynamoCore/Graph/Workspaces/PackageDependencyInfo.cs
@@ -4,19 +4,35 @@ using System.Collections.Generic;
 namespace Dynamo.Graph.Workspaces
 {
     /// <summary>
-    /// Class containing info about a package
+    /// Interface for types containing info about a package
     /// </summary>
-    public class PackageInfo
+    public interface IPackageInfo
     {
         /// <summary>
         /// Name of the package
         /// </summary>
-        internal string Name { get; set; }
+        string Name { get; }
 
         /// <summary>
         /// Version of the package
         /// </summary>
-        internal Version Version { get; set; }
+        Version Version { get; }
+    }
+
+    /// <summary>
+    /// Class containing info about a package
+    /// </summary>
+    public class PackageInfo : IPackageInfo
+    {
+        /// <summary>
+        /// Name of the package
+        /// </summary>
+        public string Name { get; }
+
+        /// <summary>
+        /// Version of the package
+        /// </summary>
+        public Version Version { get; }
 
         /// <summary>
         /// Create a package info object from the package name and version
@@ -82,7 +98,7 @@ namespace Dynamo.Graph.Workspaces
         DYFFILE
     }
 
-
+    
 
     /// <summary>
     /// An interface that describes a dependency a workspace can have on other code.
@@ -124,7 +140,7 @@ namespace Dynamo.Graph.Workspaces
     /// <summary>
     /// Class containing info about a workspace package dependency
     /// </summary>
-    internal class PackageDependencyInfo : INodeLibraryDependencyInfo
+    internal class PackageDependencyInfo : INodeLibraryDependencyInfo, IPackageInfo
     {
         /// <summary>
         /// PackageInfo for this package
