+++ src/DynamoCoreWpf/ViewModels/PackageManager/PackageManagerSearchViewModel.cs
@@ -586,180 +586,9 @@ namespace Dynamo.PackageManager
             this.SearchResults.Clear();
         }
 
-        private string JoinPackageNames(IEnumerable<Package> pkgs)
-        {
-            return String.Join(", ", pkgs.Select(x => x.Name + " " + x.VersionName));
-        } 
-
         private void PackageOnExecuted(PackageManagerSearchElement element, PackageVersion version, string downloadPath)
         {
-            string msg = String.IsNullOrEmpty(downloadPath) ?
-                String.Format(Resources.MessageConfirmToInstallPackage, element.Name, version.version) :
-                String.Format(Resources.MessageConfirmToInstallPackageToFolder, element.Name, version.version, downloadPath);
-
-            var result = MessageBox.Show(msg,
-                Resources.PackageDownloadConfirmMessageBoxTitle,
-                MessageBoxButton.OKCancel, MessageBoxImage.Question);
-
-            var pmExt = PackageManagerClientViewModel.DynamoViewModel.Model.GetPackageManagerExtension();
-            if (result == MessageBoxResult.OK)
-            {
-                // get all of the headers
-                var headers = version.full_dependency_ids.Select(dep => dep._id).Select((id) =>
-                {
-                    PackageHeader pkgHeader;
-                    var res = pmExt.PackageManagerClient.DownloadPackageHeader(id, out pkgHeader);
-
-                    if (!res.Success)
-                        MessageBox.Show(String.Format(Resources.MessageFailedToDownloadPackage, id),
-                            Resources.PackageDownloadErrorMessageBoxTitle,
-                            MessageBoxButton.OK, MessageBoxImage.Error);
-
-                    return pkgHeader;
-                }).ToList();
-
-                // if any header download fails, abort
-                if (headers.Any(x => x == null))
-                {
-                    return;
-                }
-
-                var allPackageVersions = PackageManagerSearchElement.ListRequiredPackageVersions(headers, version);
-
-                // determine if any of the packages contain binaries or python scripts.  
-                var containsBinaries =
-                    allPackageVersions.Any(
-                        x => x.Item2.contents.Contains(PackageManagerClient.PackageContainsBinariesConstant) || x.Item2.contains_binaries);
-
-                var containsPythonScripts =
-                    allPackageVersions.Any(
-                        x => x.Item2.contents.Contains(PackageManagerClient.PackageContainsPythonScriptsConstant));
-
-                // if any do, notify user and allow cancellation
-                if (containsBinaries || containsPythonScripts)
-                {
-                    var res = MessageBox.Show(Resources.MessagePackageContainPythonScript,
-                        Resources.PackageDownloadMessageBoxTitle,
-                        MessageBoxButton.OKCancel, MessageBoxImage.Exclamation);
-
-                    if (res == MessageBoxResult.Cancel) return;
-                }
-
-                // Determine if there are any dependencies that are made with a newer version
-                // of Dynamo (this includes the root package)
-                var dynamoVersion = this.PackageManagerClientViewModel.DynamoViewModel.Model.Version;
-                var dynamoVersionParsed = VersionUtilities.PartialParse(dynamoVersion, 3);
-                var futureDeps = allPackageVersions.FilterFuturePackages(dynamoVersionParsed);
-
-                // If any of the required packages use a newer version of Dynamo, show a dialog to the user
-                // allowing them to cancel the package download
-                if (futureDeps.Any())
-                {
-                    var versionList = FormatPackageVersionList(futureDeps);
-
-                    if (MessageBox.Show(String.Format(Resources.MessagePackageNewerDynamo,
-                        PackageManagerClientViewModel.DynamoViewModel.BrandingResourceProvider.ProductName,
-                        versionList),
-                        string.Format(Resources.PackageUseNewerDynamoMessageBoxTitle,
-                        PackageManagerClientViewModel.DynamoViewModel.BrandingResourceProvider.ProductName),
-                        MessageBoxButton.OKCancel,
-                        MessageBoxImage.Warning) == MessageBoxResult.Cancel)
-                    {
-                        return;
-                    }
-                }
-
-                var localPkgs = pmExt.PackageLoader.LocalPackages;
-
-                var uninstallsRequiringRestart = new List<Package>();
-                var uninstallRequiringUserModifications = new List<Package>();
-                var immediateUninstalls = new List<Package>();
-
-                // if a package is already installed we need to uninstall it, allowing
-                // the user to cancel if they do not want to uninstall the package
-                foreach (var localPkg in headers.Select(x => localPkgs.FirstOrDefault(v => v.Name == x.name)))
-                {
-                    if (localPkg == null) continue;
-
-                    if (localPkg.LoadedAssemblies.Any())
-                    {
-                        uninstallsRequiringRestart.Add(localPkg);
-                        continue;
-                    }
-
-                    if (localPkg.InUse(this.PackageManagerClientViewModel.DynamoViewModel.Model))
-                    {
-                        uninstallRequiringUserModifications.Add(localPkg);
-                        continue;
-                    }
-
-                    immediateUninstalls.Add(localPkg);
-                }
-
-                if (uninstallRequiringUserModifications.Any())
-                {
-                    MessageBox.Show(String.Format(Resources.MessageUninstallToContinue,
-                        PackageManagerClientViewModel.DynamoViewModel.BrandingResourceProvider.ProductName,
-                        JoinPackageNames(uninstallRequiringUserModifications)),
-                        Resources.CannotDownloadPackageMessageBoxTitle,
-                        MessageBoxButton.OK, MessageBoxImage.Error);
-                    return;
-                }
-
-                var settings = PackageManagerClientViewModel.DynamoViewModel.Model.PreferenceSettings;
-
-                if (uninstallsRequiringRestart.Any())
-                {
-
-                    var message = string.Format(Resources.MessageUninstallToContinue2,
-                        PackageManagerClientViewModel.DynamoViewModel.BrandingResourceProvider.ProductName,
-                        JoinPackageNames(uninstallsRequiringRestart),
-                        element.Name + " " + version.version);
-                    // different message for the case that the user is
-                    // trying to install the same package/version they already have installed.
-                    if(uninstallsRequiringRestart.Count == 1 && 
-                        uninstallsRequiringRestart.First().Name ==  element.Name &&
-                        uninstallsRequiringRestart.First().VersionName == version.version)
-                    {
-                        message = String.Format(Resources.MessageUninstallSamePackage, element.Name + " " + version.version);
-                    }
-                    var dialogResult = MessageBox.Show(message, 
-                        Resources.CannotDownloadPackageMessageBoxTitle,
-                        MessageBoxButton.YesNo, MessageBoxImage.Error);
-
-                    if (dialogResult == MessageBoxResult.Yes)
-                    {
-                        // mark for uninstallation
-                        uninstallsRequiringRestart.ForEach(x => x.MarkForUninstall(settings));
-                    }
-                    return;
-                }
-
-                if (immediateUninstalls.Any())
-                {
-                    // if the package is not in use, tell the user we will be uninstall it and give them the opportunity to cancel
-                    if (MessageBox.Show(String.Format(Resources.MessageAlreadyInstallDynamo,
-                        PackageManagerClientViewModel.DynamoViewModel.BrandingResourceProvider.ProductName,
-                        JoinPackageNames(immediateUninstalls)),
-                        Resources.DownloadWarningMessageBoxTitle,
-                        MessageBoxButton.OKCancel, MessageBoxImage.Warning) == MessageBoxResult.Cancel)
-                        return;
-                }
-
-                // add custom path to custom package folder list
-                if (!String.IsNullOrEmpty(downloadPath))
-                {
-                    if (!settings.CustomPackageFolders.Contains(downloadPath))
-                        settings.CustomPackageFolders.Add(downloadPath);
-                }
-
-                // form header version pairs and download and install all packages
-                allPackageVersions
-                        .Select(x => new PackageDownloadHandle(x.Item1, x.Item2))
-                        .ToList()
-                        .ForEach(x => this.PackageManagerClientViewModel.DownloadAndInstall(x, downloadPath));
-
-            }
+            this.PackageManagerClientViewModel.ExecutePackage(element.Name, version, downloadPath);
         }
 
         /// <summary>
