+++ src/DynamoPackages/PackageManagerClient.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Dynamo.Graph.Workspaces;
 using Greg;
 using Greg.Requests;
 using Greg.Responses;
@@ -87,6 +88,41 @@ namespace Dynamo.PackageManager
                 var pkgResponse = this.client.ExecuteAndDeserializeWithContent<List<PackageHeader>>(nv);
                 return pkgResponse.content;
             }, new List<PackageHeader>());
+        }
+
+        /// <summary>
+        /// Gets the PackageHeader for a specific package
+        /// </summary>
+        /// <param name="packageInfo"></param>
+        /// <returns></returns>
+        internal PackageHeader GetPackageHeader(IPackageInfo packageInfo)
+        {
+            var header = FailFunc.TryExecute(() =>
+            {
+                var nv = new HeaderDownload("dynamo", packageInfo.Name);
+                var pkgResponse = this.client.ExecuteAndDeserializeWithContent<PackageHeader>(nv);
+                return pkgResponse.content;
+            }, null);
+
+            return header;
+        }
+
+        /// <summary>
+        /// Gets the Greg PackageVersion object for a specific version of a package
+        /// </summary>
+        /// <param name="header"></param>
+        /// <param name="version"></param>
+        /// <returns></returns>
+        internal PackageVersion GetGregPackageVersion(PackageHeader header, Version version)
+        {
+            foreach (var v in header.versions)
+            {
+                if (new Version(v.version) == version)
+                {
+                    return v;
+                }
+            }
+            return null;
         }
 
         /// <summary>
